Refactor logging interface to have one function
llog() is now the only function in this logger, and greatly simplifies
the calling convention. In accordance with many other popular logging
libraries, the log level is not a bitwise flag, but a hard cap, such
that any log severity below LOG_CUTOFF_LEVEL is ignored. While the if
statement at the start of llog's implementation could have been
implemented in a wrapping macro, this would require the log level be
declared extern, and thus open to more race conditions when changing.
This patch does not attempt to remedy the issues with writing to a file
before locking it.