Refactor internals of `llog()` to remove switch
Greatly improve the readability of the `llog()` internals, making it
more extendible in the future also. Since the static string array in
`log.c` can easily be added to, more log levels can be added with
one line.

I would feel bad if I didn't show this trick and I forgot when sending
the last PR....